+++ studio.html
@@ -7,6 +7,14 @@
         <meta name="nofollow-links" content="on" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <meta name="aem-base-url" />
+        <script type="importmap">
+            {
+              "imports": {
+                "lit": "https://cdn.jsdelivr.net/gh/lit/dist@2.8.0/all/lit-all.min.js",
+                "lit/directives/unsafe-html.js": "https://cdn.jsdelivr.net/gh/lit/dist@2.8.0/all/lit-all.min.js"
+              }
+            }
+        </script>
         <script>
             const baseUrl = window.location.href.includes('localhost') ? 'http://localhost:8080' : 'https://author-p22655-e155390.adobeaemcloud.com';
             document.querySelector('meta[name="aem-base-url"]')?.setAttribute('content', baseUrl);
@@ -15,7 +23,7 @@
         <script src="studio/libs/maslib.js" type="module"></script>
         <script src="studio/libs/swc.js" type="module"></script>
         <script src="studio/libs/aem.js" type="module"></script>
-        <script src="studio/libs/studio.js" type="module"></script>
+        <script src="studio/src/studio.js" type="module"></script>
 
         <!-- OST -->
         <script src="studio/ost/index.js"></script>
